+++ source/views/view_strings.cpp
@@ -55,7 +55,7 @@ namespace hex {
             std::vector<u8> buffer(1024, 0x00);
             u32 foundCharacters = 0;
             for (u64 offset = 0; offset < this->m_dataProvider->getSize(); offset += buffer.size()) {
-                size_t readSize = std::min(buffer.size(), this->m_dataProvider->getSize() - offset);
+                size_t readSize = std::min(u64(buffer.size()), this->m_dataProvider->getSize() - offset);
                 this->m_dataProvider->read(offset,  buffer.data(), readSize);
 
                 for (u32 i = 0; i < readSize; i++) {
